requirements.txt
rtree==0.9.7
geopandas==0.9.0
pyyaml==5.4.1
ptitprince==0.2.5